desktop_assets/common/fredica-pyutil/requirements.txt
setuptools
wheel
loguru
fastapi[all]>=0.135.0
uvicorn[standard]
websockets
starlette
bilibili-api-python
curl_cffi
aiohttp
pydantic
psutil
pynvml
huggingface_hub
pytest
argon2-cffi
cryptography